Temporary Commit
This point I am trying to add the password supporting for the application. However, the application is heavily rely on the router URL and query: I decided to use Pinia to manage the global state (e.g. Route variables) and help improve the potential type missing occurred now in DriveList.

Note: this commit is just for recording the state without the Pinia. There are bugs left inside. Planning to solve them during the reconstruction.