Partial type correction & Building fix (Temporary)
For current stage we will dismiss the strict vue-tsc for building, as some of the function is under construction, and it may not cause running errors for this incomplete version.

The ts check are now separated to an individual run scripts.

Considering enable strict type check in IDE to avoid typing problem during development.